refactored and updated; still need to remove old/unused files
