+++ frontend/src/app/app.component.ts
@@ -54,12 +54,13 @@ export class AppComponent implements OnInit {
     // this.simulateUserLogin('test@test.com');              // Login as different user
     
     // OR navigate directly to any page for testing (bypasses auth entirely):
-    // this.router.navigate(['/investment-survey'], { replaceUrl: true });
+    this.router.navigate(['/investment-schedule'], { replaceUrl: true });
     // this.router.navigate(['/manual-stock-selection'], { replaceUrl: true });
     // this.router.navigate(['/tabs/tab1'], { replaceUrl: true });
     
+    // 🚫 COMMENT OUT AUTH LOGIC WHEN TESTING SPECIFIC PAGES
     // Subscribe to authentication state changes
-    this.authService.isLoggedIn$.subscribe(isLoggedIn => {
+    /*this.authService.isLoggedIn$.subscribe(isLoggedIn => {
       if (isLoggedIn) {
         // User is logged in, wait for progress data
         this.authService.userProgress$.subscribe(progress => {
@@ -72,7 +73,7 @@ export class AppComponent implements OnInit {
         console.log('[AppComponent] User not logged in, redirecting to get-started');
         this.router.navigate(['/get-started'], { replaceUrl: true });
       }
-    });
+    });*/
   }
 
   /**
