+++ frontend/src/app/app.routes.ts
@@ -38,6 +38,10 @@ export const routes: Routes = [
     loadComponent: () => import('./components/linkplaid/linkplaid.component').then(m => m.LinkPlaidComponent)
   },
   {
+    path: 'investment-schedule',
+    loadComponent: () => import('./components/investment-schedule/investment-schedule.component').then(m => m.InvestmentScheduleComponent)
+  },
+  {
     path: 'survey',
     loadComponent: () => import('./components/survey/survey.component').then(m => m.SurveyComponent)
     // You could add a canActivate guard here if needed later
