+++ frontend/src/app/components/investment-schedule/investment-schedule.component.old.scss
@@ -1,0 +1,293 @@
+// Use CSS variables for a consistent theme
+:host {
+  --ion-background-color: #f2f2f7; // A very light, clean background
+  --ion-toolbar-background: var(--ion-background-color);
+  --ion-text-color: #000000;
+  --ion-color-primary: #007aff;
+  --ion-color-medium: #// Responsive design
+@media (max-width: 768px) {
+  .projection-grid {
+    grid-template-columns: 1fr;
+    gap: 8px;
+  }
+  
+  .projection-item {
+    .projection-value {
+      font-size: 20px;
+    }
+  }
+
+  .section-header h2 {
+    font-size: 24px;
+  }
+}ion-toolbar {
+  --border-width: 0 !important;
+}
+
+ion-progress-bar {
+  height: 3px;
+  --progress-background: var(--ion-color-primary);
+}
+
+// Unified container for the form
+.investment-schedule-container {
+  padding: 16px;
+}
+
+.section-header {
+  margin-bottom: 32px;
+  text-align: center;
+
+  h2 {
+    font-size: 26px;
+    font-weight: 700;
+    line-height: 1.25;
+    margin: 0 0 8px 0;
+  }
+
+  .helper-text {
+    font-size: 16px;
+    color: var(--ion-color-medium);
+    margin: 0;
+    line-height: 1.4;
+  }
+}
+
+.schedule-section {
+  margin-bottom: 24px;
+}
+
+// iOS-style cards
+ion-card {
+  margin: 0 0 16px 0;
+  box-shadow: none;
+  border-radius: 10px;
+  background: #ffffff;
+  border: 1px solid #d1d1d6;
+
+  ion-card-header {
+    padding: 16px 16px 8px 16px;
+
+    ion-card-title {
+      font-size: 15px;
+      font-weight: 600;
+      color: var(--ion-text-color);
+      margin: 0;
+      text-transform: capitalize;
+    }
+  }
+
+  ion-card-content {
+    padding: 8px 16px 16px 16px;
+  }
+}
+
+// iOS-style frequency explanation
+.frequency-explanation {
+  margin-top: 12px;
+  padding: 12px;
+  background: #f2f2f7;
+  border-radius: 8px;
+
+  ion-note {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    font-size: 14px;
+    color: var(--ion-color-medium);
+  }
+}
+
+// iOS-style input styling
+.amount-input-wrapper {
+  display: flex;
+  align-items: center;
+  border: 1px solid #d1d1d6;
+  border-radius: 8px;
+  padding: 12px 16px;
+  margin-bottom: 16px;
+  background: #ffffff;
+
+  &:focus-within {
+    border-color: var(--ion-color-primary);
+    box-shadow: 0 0 0 1px var(--ion-color-primary);
+  }
+
+  .currency-symbol {
+    font-size: 17px;
+    font-weight: 500;
+    color: var(--ion-text-color);
+    margin-right: 8px;
+  }
+
+  ion-input {
+    --padding-start: 0;
+    --padding-end: 0;
+    font-size: 17px;
+    font-weight: 500;
+    color: var(--ion-text-color);
+  }
+}
+
+// iOS-style recommendation banner
+.amount-recommendations {
+  padding: 12px;
+  background: rgba(0, 122, 255, 0.1);
+  border-radius: 8px;
+  border: 1px solid rgba(0, 122, 255, 0.2);
+
+  ion-note {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    font-size: 14px;
+    color: var(--ion-color-primary);
+  }
+}
+
+// iOS-style projections
+.projections-section {
+  margin-bottom: 24px;
+}
+
+.projection-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 12px;
+  margin-bottom: 16px;
+}
+
+.projection-item {
+  text-align: center;
+  padding: 16px;
+  background: #ffffff;
+  border-radius: 8px;
+  border: 1px solid #d1d1d6;
+
+  .projection-value {
+    font-size: 22px;
+    font-weight: 600;
+    color: var(--ion-color-primary);
+    margin-bottom: 4px;
+  }
+
+  .projection-label {
+    font-size: 13px;
+    color: var(--ion-color-medium);
+    font-weight: 400;
+  }
+}
+
+// iOS-style goal comparison
+.goal-comparison {
+  padding: 12px;
+  border-radius: 8px;
+  text-align: center;
+
+  ion-note {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 8px;
+    font-size: 14px;
+    font-weight: 500;
+
+    &[color="success"] {
+      background: rgba(52, 199, 89, 0.1);
+      border: 1px solid rgba(52, 199, 89, 0.3);
+      padding: 12px;
+      border-radius: 8px;
+      color: #34c759;
+    }
+
+    &[color="warning"] {
+      background: rgba(255, 149, 0, 0.1);
+      border: 1px solid rgba(255, 149, 0, 0.3);
+      padding: 12px;
+      border-radius: 8px;
+      color: #ff9500;
+    }
+  }
+}
+
+// iOS-style summary
+.summary-section {
+  margin-top: 24px;
+
+  .summary-details {
+    margin-bottom: 16px;
+
+    p {
+      margin: 8px 0;
+      font-size: 15px;
+      line-height: 1.4;
+
+      strong {
+        color: var(--ion-text-color);
+        font-weight: 600;
+      }
+    }
+  }
+
+  ion-note {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+    font-size: 14px;
+    padding: 12px;
+    background: rgba(52, 199, 89, 0.1);
+    border-radius: 8px;
+    border: 1px solid rgba(52, 199, 89, 0.3);
+    color: #34c759;
+  }
+}
+
+// iOS-style form items
+ion-item {
+  --background: transparent;
+  --border-color: transparent;
+  --padding-start: 0;
+  --inner-padding-end: 0;
+  --min-height: 44px;
+
+  ion-select, ion-input {
+    font-size: 17px;
+    font-weight: 400;
+  }
+}
+
+// iOS-style navigation controls
+.navigation-controls {
+  margin-top: 32px;
+  padding-bottom: 16px;
+
+  ion-button {
+    --border-radius: 8px;
+    height: 50px;
+    font-size: 17px;
+    font-weight: 600;
+    margin: 0;
+
+    &[disabled] {
+      --opacity: 0.4;
+    }
+  }
+}
+
+// Responsive design
+@media (max-width: 768px) {
+  .projection-grid {
+    grid-template-columns: 1fr;
+    gap: 8px;
+  }
+  
+  .projection-item {
+    .projection-value {
+      font-size: 20px;
+    }
+  }
+
+  .section-header h2 {
+    font-size: 24px;
+  }
+}}
