+++ frontend/src/app/components/investment-schedule/investment-schedule.component.scss
@@ -1,0 +1,282 @@
+// Use CSS variables for a consistent theme
+:host {
+  --ion-background-color: #f2f2f7; // A very light, clean background
+  --ion-toolbar-background: var(--ion-background-color);
+  --ion-text-color: #000000;
+  --ion-color-primary: #007aff;
+  --ion-color-medium: #8e8e93;
+}
+
+ion-toolbar {
+  --border-width: 0 !important;
+}
+
+ion-progress-bar {
+  height: 3px;
+  --progress-background: var(--ion-color-primary);
+}
+
+// Unified container for the form
+.investment-schedule-container {
+  padding: 16px;
+}
+
+.page-intro {
+  margin-bottom: 32px;
+  text-align: center;
+
+  h2 {
+    font-size: 26px;
+    font-weight: 700;
+    line-height: 1.25;
+    margin: 0 0 8px 0;
+  }
+
+  .helper-text {
+    font-size: 16px;
+    color: var(--ion-color-medium);
+    margin: 0;
+    line-height: 1.4;
+  }
+}
+
+// Input sections styling
+.input-section {
+  margin-bottom: 32px;
+
+  .section-label {
+    margin-bottom: 16px;
+
+    h3 {
+      font-size: 20px;
+      font-weight: 600;
+      line-height: 1.25;
+      margin: 0 0 4px 0;
+      color: var(--ion-text-color);
+    }
+
+    .helper-text {
+      font-size: 14px;
+      color: var(--ion-color-medium);
+      margin: 0;
+      line-height: 1.4;
+    }
+  }
+
+  .input-display-wrapper {
+    display: flex;
+    align-items: center;
+    border-bottom: 2px solid #e5e5ea;
+    margin-bottom: 8px;
+    padding-bottom: 8px;
+
+    &:focus-within {
+      border-bottom-color: var(--ion-color-primary);
+    }
+
+    .currency-symbol {
+      font-weight: 500;
+      font-size: 18px;
+      color: var(--ion-text-color);
+      margin-right: 8px;
+    }
+
+    ion-input, ion-select {
+      --padding-start: 0;
+      --padding-end: 0;
+      font-size: 18px;
+      font-weight: 400;
+      color: var(--ion-text-color);
+    }
+
+    ion-select {
+      --placeholder-color: var(--ion-color-medium);
+    }
+  }
+
+  .amount-recommendations {
+    margin-top: 12px;
+
+    .helper-text {
+      display: flex;
+      align-items: center;
+      gap: 8px;
+      font-size: 14px;
+      color: var(--ion-color-primary);
+      margin: 0;
+
+      ion-icon {
+        font-size: 16px;
+      }
+    }
+  }
+}
+
+// Projections section
+.projections-section {
+  margin-bottom: 32px;
+
+  .section-label {
+    margin-bottom: 16px;
+
+    h3 {
+      font-size: 20px;
+      font-weight: 600;
+      line-height: 1.25;
+      margin: 0 0 4px 0;
+      color: var(--ion-text-color);
+    }
+
+    .helper-text {
+      font-size: 14px;
+      color: var(--ion-color-medium);
+      margin: 0;
+      line-height: 1.4;
+    }
+  }
+
+  .projection-grid {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 12px;
+    margin-bottom: 16px;
+  }
+
+  .projection-item {
+    text-align: center;
+    padding: 20px 16px;
+    background: #ffffff;
+    border-radius: 10px;
+    border: 1px solid #d1d1d6;
+
+    .projection-value {
+      font-size: 22px;
+      font-weight: 600;
+      color: var(--ion-color-primary);
+      margin-bottom: 4px;
+    }
+
+    .projection-label {
+      font-size: 13px;
+      color: var(--ion-color-medium);
+      font-weight: 400;
+    }
+  }
+
+  .goal-comparison {
+    .helper-text {
+      display: flex;
+      align-items: center;
+      gap: 8px;
+      font-size: 14px;
+      margin: 0;
+      line-height: 1.4;
+
+      &.success-text {
+        color: #34c759;
+      }
+
+      &.warning-text {
+        color: #ff9500;
+      }
+
+      ion-icon {
+        font-size: 16px;
+      }
+    }
+  }
+}
+
+// Summary section
+.summary-section {
+  margin-bottom: 32px;
+
+  .section-label {
+    margin-bottom: 16px;
+
+    h3 {
+      font-size: 20px;
+      font-weight: 600;
+      line-height: 1.25;
+      margin: 0 0 4px 0;
+      color: var(--ion-text-color);
+    }
+
+    .helper-text {
+      font-size: 14px;
+      color: var(--ion-color-medium);
+      margin: 0;
+      line-height: 1.4;
+    }
+  }
+
+  .summary-card {
+    background: #ffffff;
+    border-radius: 10px;
+    border: 1px solid #d1d1d6;
+    padding: 16px;
+
+    .summary-item {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      padding: 8px 0;
+
+      &:not(:last-child) {
+        border-bottom: 1px solid #f2f2f7;
+      }
+
+      .summary-label {
+        font-size: 15px;
+        color: var(--ion-color-medium);
+        font-weight: 400;
+      }
+
+      .summary-value {
+        font-size: 15px;
+        color: var(--ion-text-color);
+        font-weight: 500;
+      }
+    }
+  }
+}
+
+// Navigation controls
+.navigation-controls {
+  margin-top: 32px;
+  padding-bottom: 16px;
+
+  ion-button {
+    --border-radius: 10px;
+    height: 50px;
+    font-size: 17px;
+    font-weight: 600;
+    margin: 0;
+
+    &[disabled] {
+      --opacity: 0.4;
+    }
+  }
+}
+
+// Responsive design
+@media (max-width: 768px) {
+  .projection-grid {
+    grid-template-columns: 1fr;
+    gap: 8px;
+  }
+  
+  .projection-item {
+    .projection-value {
+      font-size: 20px;
+    }
+  }
+
+  .page-intro h2 {
+    font-size: 24px;
+  }
+
+  .section-label h3 {
+    font-size: 18px;
+  }
+}
